# Optimize FileUpdateChecker by caching mtime