<!--
Thanks for contributing to Rails!

Please do not make *Draft* pull requests, as they still send
notifications to everyone watching the Rails repo.

Create a pull request when it is ready for review and feedback
from the Rails team :).

If your pull request affects documentation or any non-code
changes, guidelines for those changes are [available
here](https://edgeguides.rubyonrails.org/contributing_to_ruby_on_rails.html#contributing-to-the-rails-documentation)

About this template

The following template aims to help contributors write a good description for their pull requests.
We'd like you to provide a description of the changes in your pull request (i.e. bugs fixed or features added), the motivation behind the changes, and complete the checklist below before opening a pull request.

Feel free to discard it if you need to (e.g. when you just fix a typo). -->

### Motivation / Background

<!--
Describe why this Pull Request needs to be merged. What bug have you fixed? What feature have you added? Why is it important?
If you are fixing a specific issue, include "Fixes #ISSUE" (replace with the issue number, remove the quotes) and the issue will be linked to this PR.
-->

This Pull Request aims to optimize `FileUpdateChecker#updated?`.


### Detail
The current implementation makes redundant system calls by checking file existence and modification time separately(
`File.exist?` & `File.mtime`)
By Using `File.mtime` and rescue `Errno::ENOENT` instead of `File.exist?`, I would like to improve its performance.
While using `rescue` typically has overhead in Ruby, in development environments, files are frequently modified, so the somewhat caching provides some benefits.

Our benchmark is;
```
ruby 3.3.7 (2025-01-15 revision be31f993d7) [arm64-darwin24]
Warming up --------------------------------------
       without cache     1.000 i/100ms
          with cache     1.000 i/100ms
Calculating -------------------------------------
       without cache     11.512 (± 8.7%) i/s   (86.86 ms/i) -    112.000 in  10.012701s
          with cache     16.313 (± 6.1%) i/s   (61.30 ms/i) -    163.000 in  10.025499s

Comparison:
          with cache:       16.3 i/s
       without cache:       11.5 i/s - 1.42x  slower
```

### Additional information

<!-- Provide additional information such as benchmarks, references to other repositories, or alternative solutions. -->

### Checklist

Before submitting the PR make sure the following are checked:

* [ ] This Pull Request is related to one change. Unrelated changes should be opened in separate PRs.
* [ ] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [ ] Tests are added or updated if you fix a bug or add a feature.
* [ ] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.